Include anti-CSRF middleware in PZH Web Server
The PZH Web Server was horribly vulnerable to CSRF attacks, particularly
as it offers so many powerful POST interfaces and relies on a single
log-in system.

This commit fixes the problem by using the connect csrf middleware, and
refactoring a lot of the code to be two-stage rather than one-stage
events.

In addition, the main page has been refactored into a number of templates
for ease of use, and to enhance readability.

Jira issue: WP-781