Add limited support for facebook invites and generic invite links
This commit makes it possible for every user to generate a link to
their own PZH which, if followed by a trusted user, will then start
the connection process.

This can be sent over Twitter or (with this PR) via a Facebook message.

In addition, support has been added for Google URL shortening of link
messages, to make everything a bit more usable.

Next step is to auto-fill the PZH details of the end user.

Jira issue: WP-852