Bug fixes for invitations and obtaining PZH info from PZP
This patch requests the user's PZH address from their PZP when
an invitation is received.  It will then auto-fill the right PZP
address making the whole process much easier.

It also solves some of the redirection issues and updates
the configuration to allow connections to other URIs.

Jira issue: WP-852